Refactoring: Remaining changes
