+++ application/backend/app/lifecycle.py
@@ -17,8 +17,8 @@ from app.core.jobs import JobController, JobQueue, ProcessRunnerFactory
 from app.core.logging import LogConfig, setup_logging, setup_uvicorn_logging
 from app.core.run import Runnable, RunnableFactory
 from app.db import MigrationManager, get_db_session
+from app.models.job import JobType
 from app.scheduler import Scheduler
-from app.schemas.job import JobType
 from app.services import (
     DatasetRevisionService,
     DatasetService,
