+++ application/backend/app/models/model_architecture.py
