--- application/backend/app/schemas/__init__.py
@@ -1,13 +1,0 @@
-# Copyright (C) 2025 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-
-from app.schemas.metrics import InferenceMetrics, LatencyMetrics, PipelineMetrics, TimeWindow
-from app.schemas.model_architecture import ModelArchitectures
-
-__all__ = [
-    "InferenceMetrics",
-    "LatencyMetrics",
-    "ModelArchitectures",
-    "PipelineMetrics",
-    "TimeWindow",
-]
