+++ application/backend/app/services/pipeline_metrics_service.py
@@ -6,7 +6,7 @@ from datetime import UTC, datetime, timedelta
 from uuid import UUID
 
 from app.models import PipelineStatus
-from app.schemas.metrics import InferenceMetrics, LatencyMetrics, PipelineMetrics, ThroughputMetrics, TimeWindow
+from app.models.metrics import InferenceMetrics, LatencyMetrics, PipelineMetrics, ThroughputMetrics, TimeWindow
 
 from .metrics_service import MetricsService
 from .pipeline_service import PipelineService
